Keep trying Android reset
